Refactor tools into category-based folder structure
- Create tools/ folder with one file per category:
  - dom.py: ClickTool, FillTool, TypeTool, UploadTool
  - navigation.py: GotoTool
  - tab.py: OpenTabTool, SwitchTabTool
  - utility.py: WaitTool
  - control.py: CompleteWorkTool, AbortWorkTool
  - pixel.py: placeholder for Computer Use tools
- Remove old tools.py file
- Prepares for adding pixel-based tools for Computer Use integration
