+++ src/webtask/_internal/agent/tools/pixel.py
@@ -1,0 +1,4 @@
+"""Pixel-based tools that use screen coordinates."""
+
+# Placeholder for Computer Use tools
+# Will contain: ClickAtTool, HoverAtTool, TypeAtTool, ScrollAtTool, DragAndDropTool, etc.
